backend/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alabaster==1.0.0
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.35
bcrypt==4.3.0
beautifulsoup4==4.8.2
blinker==1.9.0
brotli
cachetools==5.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
compressed-rtf==1.0.7
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
docutils==0.21.2
docx2txt==0.9
ebcdic==1.1.1
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
extract-msg==0.28.7
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.5.0
ftfy==6.3.1
future==1.0.0
gmpy2
GPUtil==1.4.0
greenlet==3.2.1
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
hyperframe==5.2.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
IMAPClient==2.1.0
imgaug==0.4.0
indic-nlp-library-itt==0.1.0
-e git+https://github.com/VarunGumma/IndicTransToolkit.git@728a7a9e8bcbbc59ca9e15c4297889df866b8c4a#egg=IndicTransToolkit
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.147
language_tool_python==2.9.3
lazy_loader==0.4
lmdb==1.6.2
lxml==5.4.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
mkl-service==2.4.0
mkl_fft
mkl_random
more-itertools==10.7.0
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.3
nltk==3.9.1
numpy==1.24.4
olefile==0.47
openai==1.63.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
packaging==24.2
paddleocr==2.6.1.3
paddlepaddle==2.5.2
pandas==1.5.3
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20191110
pillow==11.2.1
portalocker==3.1.1
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.1
protobuf==3.20.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.20.2
pyparsing==3.2.3
pypdfium2==4.30.1
PySocks
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
pytz==2025.2
pywin32==310
PyYAML==6.0.2
RapidFuzz==3.13.0
rarfile==4.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
shapely==2.1.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
spello==1.3.0
Sphinx==8.1.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.40
stringzilla==3.12.5
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.1.0
textract==1.6.5
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.5.1+cpu
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.52.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.1.3
win-inet-pton
xlrd==1.2.0
XlsxWriter==3.2.3
yarl==1.20.0
